refactor: stash tab overlay stuff renamed, got rid of redundant code, more hardening for supporting stash tab prefix query mode
